// ui/src/data/decoders/generated/v2/ocap_v2.ts
export enum Side {
  SIDE_UNKNOWN = 0,
  SIDE_WEST = 1,
  SIDE_EAST = 2,
  SIDE_GUER = 3,
  SIDE_CIV = 4,
  SIDE_GLOBAL = 5,
  UNRECOGNIZED = -1,
}
// ...
export interface Manifest {
  /** = 2 */
  version: number;
  world: WorldMeta | undefined;
  mission: MissionMeta | undefined;
  frameCount: number;
  chunkSize: number;
  captureDelayMs: number;
  chunkCount: number;
  soldiers: SoldierDef[];
  vehicles: VehicleDef[];
  events: Event[];
  markers: MarkerDef[];
  times: TimeSample[];
}

export interface WorldMeta {
  worldName: string;
  worldSize: number;
  latitude: number;
  longitude: number;
  author: string;
  displayName: string;
}

export interface MissionMeta {
  missionName: string;
  briefingName: string;
  author: string;
  serverName: string;
  extensionVersion: string;
  addonVersion: string;
  extensionBuild: string;
  tag: string;
  playableSlots: PlayableSlots | undefined;
  sideFriendly: SideFriendly | undefined;
  addons: Addon[];
}

export interface PlayableSlots {
  west: number;
  east: number;
  independent: number;
  civilian: number;
}

export interface SideFriendly {
  eastWest: boolean;
  eastIndependent: boolean;
  westIndependent: boolean;
}

export interface Addon {
  name: string;
  workshopId: string;
}

export interface SoldierDef {
  id: number;
  name: string;
  side: Side;
  groupName: string;
  role: string;
  startFrame: number;
  endFrame: number;
  isPlayer: boolean;
  className: string;
  playerUid: string;
  framesFired: FiredFrame[];
}

export interface VehicleDef {
  id: number;
  name: string;
  /** ocapType: car, tank, heli, etc. */
  vehicleClass: string;
  className: string;
  startFrame: number;
  endFrame: number;
  customization: string;
}

export interface FiredFrame {
  frameNum: number;
  startPos: Position3D | undefined;
  endPos: Position3D | undefined;
  weapon: string;
  magazine: string;
  firingMode: string;
}

export interface Event {
  frameNum: number;
  kill?: KillEvent | undefined;
  hit?: HitEvent | undefined;
  projectile?: ProjectileEvent | undefined;
  chat?: ChatEvent | undefined;
  radio?: RadioEvent | undefined;
  ace3Death?: Ace3DeathEvent | undefined;
  ace3Unconscious?: Ace3UnconsciousEvent | undefined;
  general?: GeneralEvent | undefined;
  connect?: ConnectEvent | undefined;
  endMission?: EndMissionEvent | undefined;
  telemetry?: TelemetryEvent | undefined;
}

export interface KillEvent {
  victimSoldierId: number;
  victimVehicleId: number;
  killerSoldierId: number;
  killerVehicleId: number;
  weaponName: string;
  weaponMagazine: string;
  eventText: string;
  distance: number;
  victimIsVehicle: boolean;
  killerIsVehicle: boolean;
}

export interface HitEvent {
  victimSoldierId: number;
  victimVehicleId: number;
  shooterSoldierId: number;
  shooterVehicleId: number;
  weaponName: string;
  weaponMagazine: string;
  eventText: string;
  distance: number;
  victimIsVehicle: boolean;
  shooterIsVehicle: boolean;
}

export interface ProjectileEvent {
  firerId: number;
  vehicleId: number;
  /** WeaponDisplay */
  weapon: string;
  /** MagazineDisplay */
  magazine: string;
  simulationType: string;
  trajectory: TrajectoryPoint[];
  hits: ProjectileHit[];
  /** MuzzleDisplay */
  muzzle: string;
  /** MagazineIcon (Arma path for ammo icon) */
  magazineIcon: string;
}
// ...
export interface Chunk {
  index: number;
  startFrame: number;
  frameCount: number;
  frames: Frame[];
}

export interface Frame {
  frameNum: number;
  soldiers: SoldierState[];
  vehicles: VehicleState[];
}
// ...
export interface MarkerDef {
  type: string;
  text: string;
  startFrame: number;
  endFrame: number;
  playerId: number;
  color: string;
  side: Side;
  positions: MarkerPosition[];
  size: number[];
  shape: string;
  brush: string;
}

export interface MarkerPosition {
  frameNum: number;
  position: Position3D | undefined;
  direction: number;
  alpha: number;
  lineCoords: number[];
}

export interface TimeSample {
  frameNum: number;
  systemTimeUtc: string;
  date: string;
  timeMultiplier: number;
  time: number;
}
// ...
export const WorldMeta: MessageFns<WorldMeta> = {
  encode(message: WorldMeta, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.worldName !== "") {
      writer.uint32(10).string(message.worldName);
    }
    if (message.worldSize !== 0) {
      writer.uint32(21).float(message.worldSize);
    }
    if (message.latitude !== 0) {
      writer.uint32(29).float(message.latitude);
    }
    if (message.longitude !== 0) {
      writer.uint32(37).float(message.longitude);
    }
    if (message.author !== "") {
      writer.uint32(42).string(message.author);
    }
    if (message.displayName !== "") {
      writer.uint32(50).string(message.displayName);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): WorldMeta {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseWorldMeta();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.worldName = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 21) {
            break;
          }

          message.worldSize = reader.float();
          continue;
        }
        case 3: {
          if (tag !== 29) {
            break;
          }

          message.latitude = reader.float();
          continue;
        }
        case 4: {
          if (tag !== 37) {
            break;
          }

          message.longitude = reader.float();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.author = reader.string();
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.displayName = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): WorldMeta {
    return {
      worldName: isSet(object.worldName)
        ? globalThis.String(object.worldName)
        : isSet(object.world_name)
        ? globalThis.String(object.world_name)
        : "",
      worldSize: isSet(object.worldSize)
        ? globalThis.Number(object.worldSize)
        : isSet(object.world_size)
        ? globalThis.Number(object.world_size)
        : 0,
      latitude: isSet(object.latitude) ? globalThis.Number(object.latitude) : 0,
      longitude: isSet(object.longitude) ? globalThis.Number(object.longitude) : 0,
      author: isSet(object.author) ? globalThis.String(object.author) : "",
      displayName: isSet(object.displayName)
        ? globalThis.String(object.displayName)
        : isSet(object.display_name)
        ? globalThis.String(object.display_name)
        : "",
    };
  },

  toJSON(message: WorldMeta): unknown {
    const obj: any = {};
    if (message.worldName !== "") {
      obj.worldName = message.worldName;
    }
    if (message.worldSize !== 0) {
      obj.worldSize = message.worldSize;
    }
    if (message.latitude !== 0) {
      obj.latitude = message.latitude;
    }
    if (message.longitude !== 0) {
      obj.longitude = message.longitude;
    }
    if (message.author !== "") {
      obj.author = message.author;
    }
    if (message.displayName !== "") {
      obj.displayName = message.displayName;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<WorldMeta>, I>>(base?: I): WorldMeta {
    return WorldMeta.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<WorldMeta>, I>>(object: I): WorldMeta {
    const message = createBaseWorldMeta();
    message.worldName = object.worldName ?? "";
    message.worldSize = object.worldSize ?? 0;
    message.latitude = object.latitude ?? 0;
    message.longitude = object.longitude ?? 0;
    message.author = object.author ?? "";
    message.displayName = object.displayName ?? "";
    return message;
  },
};
// ...
export const Chunk: MessageFns<Chunk> = {
  encode(message: Chunk, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.index !== 0) {
      writer.uint32(8).uint32(message.index);
    }
    if (message.startFrame !== 0) {
      writer.uint32(16).uint32(message.startFrame);
    }
    if (message.frameCount !== 0) {
      writer.uint32(24).uint32(message.frameCount);
    }
    for (const v of message.frames) {
      Frame.encode(v!, writer.uint32(34).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Chunk {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseChunk();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.index = reader.uint32();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.startFrame = reader.uint32();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.frameCount = reader.uint32();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.frames.push(Frame.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Chunk {
    return {
      index: isSet(object.index) ? globalThis.Number(object.index) : 0,
      startFrame: isSet(object.startFrame)
        ? globalThis.Number(object.startFrame)
        : isSet(object.start_frame)
        ? globalThis.Number(object.start_frame)
        : 0,
      frameCount: isSet(object.frameCount)
        ? globalThis.Number(object.frameCount)
        : isSet(object.frame_count)
        ? globalThis.Number(object.frame_count)
        : 0,
      frames: globalThis.Array.isArray(object?.frames) ? object.frames.map((e: any) => Frame.fromJSON(e)) : [],
    };
  },

  toJSON(message: Chunk): unknown {
    const obj: any = {};
    if (message.index !== 0) {
      obj.index = Math.round(message.index);
    }
    if (message.startFrame !== 0) {
      obj.startFrame = Math.round(message.startFrame);
    }
    if (message.frameCount !== 0) {
      obj.frameCount = Math.round(message.frameCount);
    }
    if (message.frames?.length) {
      obj.frames = message.frames.map((e) => Frame.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Chunk>, I>>(base?: I): Chunk {
    return Chunk.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Chunk>, I>>(object: I): Chunk {
    const message = createBaseChunk();
    message.index = object.index ?? 0;
    message.startFrame = object.startFrame ?? 0;
    message.frameCount = object.frameCount ?? 0;
    message.frames = object.frames?.map((e) => Frame.fromPartial(e)) || [];
    return message;
  },
};
// ...
export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}